Create README.md
We created interactive simulations and animations of the Lorenz, Rossler, and Henon attractors by solving their differential equations using Euler's and Modified Euler's methods. We then compared the extent of the "butterfly effect" and noted that changes of <0.5 do not yield significant changes in the resulting figures. We also examined the 5-line fractal structure of the Henon attractor. An in-depth paper is attached which includes more details about our project.